+++ scripts/pdf-editable.mjs
@@ -156,6 +156,10 @@ Hooks.on('deleteItem',  delete_document)
 
 /**
  * Called from renderJournalPDFPageSheet
+ * @param {JournalPDFPageSheet} sheet
+ * @param {jQuery} html The iframe for the PDF Page
+ * @param {Object} data The information from sheet.getData()
+ * @param {String} id_to_display The UUID of the Actor or Item to be displayed in the Form Fillable PDF
  * @inheritData renderJournalPDFPageSheet
  */
 export async function initEditor(sheet, html, data, id_to_display) {